move generic functions to support/plex_media (cherry picked from commit 6dd87e7)
